+++ src/main/java/com/real/backend/common/util/CookieUtils.java
@@ -12,6 +12,10 @@ import jakarta.servlet.http.HttpServletResponse;
 
 @Component
 public class CookieUtils {
+
+    private final boolean isSecure = true;
+    private final boolean isHttpOnly = true;
+
     public String resolveTokenFromCookie(HttpServletRequest request, String type) {
         Cookie[] cookies = request.getCookies();
         if (cookies == null) return null;
@@ -23,10 +27,7 @@ public class CookieUtils {
             .orElse(null);
     }
 
-    public void setTokenCookie(HttpServletResponse response, String accessToken, String refreshToken) {
-        boolean isSecure = true;
-        boolean isHttpOnly = true;
-
+    public void setTokenCookies(HttpServletResponse response, String accessToken, String refreshToken) {
         ResponseCookie accessCookie = createResponseCookie("ACCESS_TOKEN", accessToken, isHttpOnly, isSecure, "/",
             "Lax");
         ResponseCookie refreshCookie = createResponseCookie("REFRESH_TOKEN", refreshToken, isHttpOnly, isSecure,
@@ -34,6 +35,17 @@ public class CookieUtils {
 
         response.addHeader(HttpHeaders.SET_COOKIE, accessCookie.toString());
         response.addHeader(HttpHeaders.SET_COOKIE, refreshCookie.toString());
+    }
+
+    public void deleteTokenCookies(HttpServletResponse response) {
+        ResponseCookie deleteAccessCookie = deleteResponseCookie("ACCESS_TOKEN",  isHttpOnly, isSecure, "/",
+            "Lax");
+        ResponseCookie deleteRefreshCookie = deleteResponseCookie("REFRESH_TOKEN",  isHttpOnly, isSecure,
+            "/api/v1/auth", "None");
+
+        response.addHeader(HttpHeaders.SET_COOKIE, deleteAccessCookie.toString());
+        response.addHeader(HttpHeaders.SET_COOKIE, deleteRefreshCookie.toString());
+
     }
 
     public ResponseCookie createResponseCookie(String name, String token, boolean isHttpOnly, boolean isSecure, String path, String sameSite) {
