+++ src/main/java/com/real/backend/modules/auth/controller/AuthController.java
@@ -8,11 +8,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.real.backend.common.response.StatusResponse;
 import com.real.backend.common.util.CookieUtils;
 import com.real.backend.modules.auth.dto.TokenDTO;
 import com.real.backend.modules.auth.kakao.KakaoUtil;
-import com.real.backend.modules.auth.service.AuthService;
-import com.real.backend.common.response.StatusResponse;
+import com.real.backend.modules.auth.service.RefreshTokenService;
+import com.real.backend.modules.auth.service.TokenService;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -22,10 +23,10 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequestMapping("/api")
 public class AuthController {
-
-    private final AuthService authService;
     private final KakaoUtil kakaoUtil;
     private final CookieUtils cookieUtils;
+    private final TokenService tokenService;
+    private final RefreshTokenService refreshTokenService;
 
     @GetMapping("/v1/oauth/{provider}")
     public void oauthLogin(@PathVariable("provider") String provider, HttpServletResponse response) throws IOException {
@@ -36,14 +37,15 @@ public class AuthController {
 
     @PostMapping("/v1/auth/logout")
     public StatusResponse logout(HttpServletRequest request, HttpServletResponse response) {
-        authService.logout(request, response);
+        refreshTokenService.deleteRefreshToken(cookieUtils.resolveTokenFromCookie(request, "REFRESH_TOKEN"));
+        cookieUtils.deleteTokenCookies(response);
         return StatusResponse.of(204, "성공적으로 로그아웃 됐습니다.");
     }
 
     @PostMapping("v1/auth/refresh")
     public StatusResponse refresh(HttpServletRequest request, HttpServletResponse response) {
-        TokenDTO tokenDTO = authService.refreshAccessToken(cookieUtils.resolveTokenFromCookie(request, "REFRESH_TOKEN"));
-        cookieUtils.setTokenCookie(response, tokenDTO.accessToken(), tokenDTO.refreshToken());
+        TokenDTO tokenDTO = tokenService.refreshAccessToken(cookieUtils.resolveTokenFromCookie(request, "REFRESH_TOKEN"));
+        cookieUtils.setTokenCookies(response, tokenDTO.accessToken(), tokenDTO.refreshToken());
         return StatusResponse.of(200, "OK");
     }
 }
