+++ src/main/java/com/real/backend/modules/auth/service/TokenService.java
@@ -1,14 +1,13 @@
 package com.real.backend.modules.auth.service;
 
-import java.time.Duration;
-import java.time.LocalDateTime;
-
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.real.backend.common.util.CONSTANT;
+import com.real.backend.common.exception.UnauthorizedException;
 import com.real.backend.modules.auth.domain.RefreshToken;
+import com.real.backend.modules.auth.dto.TokenDTO;
 import com.real.backend.modules.auth.repository.RefreshTokenRepository;
+import com.real.backend.modules.user.component.UserFinder;
 import com.real.backend.modules.user.domain.User;
 import com.real.backend.security.JwtUtils;
 
@@ -17,25 +16,26 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class TokenService {
+    private final RefreshTokenService refreshTokenService;
     private final RefreshTokenRepository refreshTokenRepository;
+    private final AccessTokenService accessTokenService;
     private final JwtUtils jwtUtils;
+    private final UserFinder userFinder;
 
     @Transactional
-    public String generateRefreshToken(User user) {
-        String refreshToken = jwtUtils.generateToken("refresh", user.getId(), user.getNickname(), user.getRole().toString(),
-                    CONSTANT.REFRESH_TOKEN_EXPIRED);
-        LocalDateTime expiryTime = LocalDateTime.now().plus(Duration.ofSeconds(CONSTANT.REFRESH_TOKEN_EXPIRED));
-        refreshTokenRepository.deleteByUser(user);
-        refreshTokenRepository.save(RefreshToken.builder()
-            .token(refreshToken)
-            .user(user)
-            .expiryTime(expiryTime)
-            .build());
-        return refreshToken;
-    }
+    public TokenDTO refreshAccessToken(String refreshToken) {
+        if (refreshToken == null) { throw new UnauthorizedException("refresh token is null"); }
 
-    public String generateAccessToken(User user) {
-        return jwtUtils.generateToken("access", user.getId(), user.getNickname(), user.getRole().toString(),
-            CONSTANT.ACCESS_TOKEN_EXPIRED);
+        User user = userFinder.getUser(jwtUtils.getId(refreshToken));
+        RefreshToken saved = refreshTokenRepository.findByUser(user).orElseThrow(() -> new UnauthorizedException("refresh token not found"));
+
+        if (!saved.getToken().equals(refreshToken)) {
+            throw new UnauthorizedException("refresh token does not match");
+        }
+
+        String accessToken = accessTokenService.generateAccessToken(user);
+        String newRefreshToken = refreshTokenService.generateRefreshToken(user);
+
+        return new TokenDTO(accessToken, newRefreshToken);
     }
 }
